EpochVersion in dev node (#2918)
* refactor: stake in contract for use in test

* dev-node: PoS version with funded stake table

- Use the epoch version in the espresso dev node

There are issues with the deploy code for the dev node. But I don't want
to rewrite more than necessary until our alloy migration PR is merged.
Currently we deploy the token and stake table also on Alt chains if
those are requested which doesn't really make sense.

* dev-node: set non-marketplace builder port

* dev-node: revert base fee to zero

The builder hardcodes a fee offering of 123 which is only sufficient if
the base fee is zero.

* Fix building docker images locally

Build in docker to avoid libssl version mismatch in nix env vs. ubuntu

* dev-node: fix epoch progression

- Make epoch height configurable

* dev-node: finalize blocks immediately

With slots_in_epoch=1 we still lag behind a block and not have the
second delegation L1 tx included in the stake table events we fetch.

* WIP: merge

* dev-node: fixup after alloy merge

* dev-node: some code cleanup

* dev-node: use l1_contracts for clarity

* comment out txn_submit check in dev node

* Real builder in devnode

* fix: state-cert URL

* clippy

* remove unnecessary /v1

* fix: test_chain_config_from_instance

Use a default max_block_size=300 if TestNetwork can't resolve chain
config.

---------

Co-authored-by: Jeremy <yiliang.he@qq.com>
Co-authored-by: Artemii Gerasimovich <artemii@espressosys.com>