L1Client: prepend latest block to block stream (#3007)
When watching blocks we only subscribe to new blocks so if the L1 does not produce any blocks we never update the state of the L1 client. For testing we may want to use anvil with on demand mining, this ensures that our L1 client quickly has blocks to work with. I think as part of a larger refactor we should make the L1 finalized non-optional and fetch the latest L1 finalized block on startup.

  * test: L1 client initialization
  * L1Client: prepend latest block to block stream
  * cache finalized block in update_loop
  * Rename some confusingly similarly named functions.
  * Use `put_finalized` method (Since we have a method for this we might as well use it.)