Eliminate linear scaling in finalized block fetching (#3131)
Eliminates scaling linear in the length of the parent chain for
fetching of finalized blocks (whether by number or timestamp). This
is particularly relevant for fetching the L1 genesis block on
startup, and in the worst case it could prevent a node from starting
up in a feasible amount of time. In both cases, fetching an old block
is expensive because the implementation works backwards block by
block from the latest block (when fetching by number, to verify hash
chains; when fetching by timestamp, to find the earliest L1 block with
a sufficient timestamp).

This change should not affect consensus much if at all, because in
consensus we generally only fetch very recent blocks for which this
working backwards doesn't cost much.

The solution is different depending on the method:
* when fetching by number, add a "safety margin"; blocks older than
  this margin are assumed to be so safe from reorg, even for lagging
  providers, that we can just fetch them without fetching the entire
  chain
* when fetching by timestamp, use binary search instead of linear
  search to find the earliest block with sufficient timestamp